Re-add fixture4
I suspect that I commited a fixture with a lock. Every time a test runs, bolt
finds that the process that had the lock does not exist anymore and removes
it. This results in a modification of the file. This simply recommits it in a
state where it should _not_ have a lock anymore.